Relocate some test files
